# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Bowtie API
  version: 0.1.0
  description: Live TV streaming API for HDHomeRun tuners.
paths:
...
  /api/v1/recordings:
    get:
      summary: List recordings (everyone's)
      operationId: listRecordings
      tags: [dvr]
      security:
        - bearerAuth: []
      parameters:
        - name: state
          in: query
          description: upcoming (scheduled/waiting/recording, oldest first), recorded (converting/ready, newest first) or failed (newest first). Omit for all.
          schema:
            type: string
            enum: [upcoming, recorded, failed]
      responses:
        '200':
          description: Recordings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Recording'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Schedule a recording
      description: |
        Either a guide program (`channelId` + `programStart`, the program's
        exact start) or a manual window (`channelId` + `start` + `stop`,
        optional `title`). Capture starts 1 min early and ends 3 min late.
        Overlapping recordings on the same channel share a tuner; more
        distinct channels than tuners is a 409 unless `force`.
      operationId: createRecording
      tags: [dvr]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [channelId]
              properties:
                channelId:
                  type: integer
                  format: int64
                programStart:
                  type: string
                  format: date-time
                start:
                  type: string
                  format: date-time
                stop:
                  type: string
                  format: date-time
                title:
                  type: string
                force:
                  type: boolean
      responses:
        '201':
          description: Scheduled
          content:
            application/json:
              schema:
                type: object
                required: [recording, warnings]
                properties:
                  recording:
                    $ref: '#/components/schemas/Recording'
                  warnings:
                    type: array
                    items:
                      type: object
                      required: [code, message]
                      properties:
                        code:
                          type: string
                          enum: [usesAllTuners]
                        message:
                          type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Unknown/disabled channel, or no such program in the guide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Tuner conflict
          content:
            application/json:
              schema:
                type: object
                required: [error, tunerCount, conflicts]
                properties:
                  error:
                    type: string
                  tunerCount:
                    type: integer
                  conflicts:
                    type: array
                    items:
                      $ref: '#/components/schemas/Recording'
        '503':
          description: Recording is not available on this server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/recordings/{id}/position:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    put:
      summary: Save the caller's resume position
      operationId: setRecordingPosition
      tags: [dvr]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [positionSec]
              properties:
                positionSec:
                  type: integer
                  minimum: 0
      responses:
        '204':
          description: Saved
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Recording:
      type: object
      required: [id, title, subtitle, description, category, channelId, channelName, start, stop, state, partial, failure, failureDetail, durationSec, sizeBytes, protected, positionSec, scheduledBy, canManage]
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        subtitle:
          type: string
        description:
          type: string
        category:
          type: string
        channelId:
          type: integer
          format: int64
        channelName:
          type: string
        start:
          type: string
          format: date-time
        stop:
          type: string
          format: date-time
        state:
          type: string
          enum: [scheduled, waiting, recording, converting, ready, failed]
          description: waiting = inside its window, retrying for a free tuner
        partial:
          type: boolean
          description: More than a minute is missing (late start, dropped stream, or a restart)
        failure:
          type: string
          enum: ['', noTuner, noSignal, diskFull, error, skipped]
        failureDetail:
          type: string
        durationSec:
          type: integer
        sizeBytes:
          type: integer
          format: int64
        protected:
          type: boolean
          description: Never deleted automatically when space runs low
        positionSec:
          type: integer
          description: The caller's resume position
        scheduledBy:
          type: string
        canManage:
          type: boolean
          description: The caller may stop, delete or protect it (scheduler or admin)
        rating:
          type: string
          description: The program's rating when scheduled ("" = not rated)
        ruleId:
          type: integer
          format: int64
          description: Series rule that scheduled it (0 = one-off)
        locked:
          type: boolean
          description: Parental controls block it for the caller (no description; play is 403)
...
  responses:
    BadRequest:
      description: Invalid request body
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication failed or missing
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'